PREAPPS-502: Add defaultProp for `sortBy` to guard against calls to `props.sortBy.indexOf`. Fixes an exception thrown when rendering the MailList with no value for `sortBy`.
